Revert claude client arch change
