+++ crates/executors/src/executors/claude/client.rs
@@ -1,4 +1,4 @@
-use std::sync::{Arc, OnceLock};
+use std::sync::Arc;
 
 use workspace_utils::approvals::ApprovalStatus;
 
@@ -9,10 +9,9 @@ use crate::{
         ExecutorError,
         claude::{
             ClaudeJson,
-            protocol::ProtocolPeer,
             types::{
-                Message, PermissionResult, PermissionUpdate, PermissionUpdateDestination,
-                PermissionUpdateType, SDKControlRequestType,
+                PermissionResult, PermissionUpdate, PermissionUpdateDestination,
+                PermissionUpdateType,
             },
         },
         codex::client::LogWriter,
@@ -23,7 +22,6 @@ const EXIT_PLAN_MODE_NAME: &str = "ExitPlanMode";
 
 /// Claude Agent client with control protocol support
 pub struct ClaudeAgentClient {
-    rpc: OnceLock<ProtocolPeer>,
     log_writer: LogWriter,
     approvals: Option<Arc<dyn ExecutorApprovalService>>,
     auto_approve: bool, // true when approvals is None
@@ -37,19 +35,10 @@ impl ClaudeAgentClient {
     ) -> Arc<Self> {
         let auto_approve = approvals.is_none();
         Arc::new(Self {
-            rpc: OnceLock::new(),
             log_writer,
             approvals,
             auto_approve,
         })
-    }
-
-    pub fn connect(&self, peer: ProtocolPeer) {
-        let _ = self.rpc.set(peer);
-    }
-
-    fn rpc(&self) -> &ProtocolPeer {
-        self.rpc.get().expect("Claude RPC peer not attached")
     }
 
     async fn handle_approval(
@@ -181,27 +170,5 @@ impl ClaudeAgentClient {
     pub async fn on_non_control(&self, line: &str) -> Result<(), ExecutorError> {
         // Forward all non-control messages to stdout
         self.log_writer.log_raw(line).await
-    }
-
-    pub async fn send_user_message(&self, content: String) -> Result<(), ExecutorError> {
-        let message = Message::new_user(content);
-        self.rpc().send_message(message).await
-    }
-
-    pub async fn initialize(&self, hooks: Option<serde_json::Value>) -> Result<(), ExecutorError> {
-        self.rpc()
-            .send_control_request(SDKControlRequestType::Initialize { hooks })
-            .await
-    }
-    pub async fn interrupt(&self) -> Result<(), ExecutorError> {
-        self.rpc()
-            .send_control_request(SDKControlRequestType::Interrupt {})
-            .await
-    }
-
-    pub async fn set_permission_mode(&self, mode: PermissionMode) -> Result<(), ExecutorError> {
-        self.rpc()
-            .send_control_request(SDKControlRequestType::SetPermissionMode { mode })
-            .await
     }
 }
